requirements.txt
gradio
PyPDF2
sentence-transformers
faiss-cpu
groq
langchain
python-dotenv